requirements.txt
torch
transformers
tqdm
numpy
wandb
python-dotenv
datasets
matplotlib